@@ -26,7 +26,7 @@
   ],
   "scripts": {
     "test": "jest --config ./jest.config.js",
-    "validate": "yarn prettier:check lint:eslint && yarn lint:css && yarn typecheck",
+    "validate": "yarn prettier:check && yarn lint:eslint && yarn lint:css && yarn typecheck",
     "prettier:check": "yarn prettier --debug-check src/**/*.{js,jsx} test/**/*.{js,jsx} stories/**/*.{js,jsx} website/src/**/*.{js,jsx}",
     "prettier:write": "yarn prettier --write src/**/*.{js,jsx} test/**/*.{js,jsx} stories/**/*.{js,jsx} website/src/**/*.{js,jsx}",
     "lint:eslint": "yarn eslint .",
@@ -57,12 +57,12 @@
     "tiny-invariant": "^0.0.3"
   },
   "devDependencies": {
-    "@atlaskit/css-reset": "^2.0.2",
-    "@storybook/react": "^3.4.6",
+    "@atlaskit/css-reset": "^2.0.6",
+    "@storybook/react": "^3.4.8",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.3",
-    "babel-jest": "^23.0.1",
+    "babel-eslint": "^8.2.5",
+    "babel-jest": "^23.2.0",
     "babel-plugin-styled-components": "^1.5.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
@@ -72,38 +72,38 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
-    "cross-env": "^5.1.6",
+    "cross-env": "^5.2.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.19.1",
+    "eslint": "^5.0.1",
     "eslint-config-airbnb": "^17.0.0",
     "eslint-config-prettier": "^2.9.0",
     "eslint-plugin-flowtype": "^2.49.3",
-    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-jest": "^21.17.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-prettier": "^2.6.1",
-    "eslint-plugin-react": "^7.9.1",
-    "flow-bin": "0.73.0",
-    "jest": "^23.1.0",
+    "eslint-plugin-react": "^7.10.0",
+    "flow-bin": "0.75.0",
+    "jest": "^23.2.0",
     "jest-watch-typeahead": "^0.1.0",
     "prettier": "^1.13.6",
-    "puppeteer": "^1.4.0",
+    "puppeteer": "^1.5.0",
     "raf-stub": "^2.0.2",
     "react": "^16.3.1",
     "react-dom": "^16.3.1",
     "react-test-renderer": "^16.3.1",
     "rimraf": "^2.6.2",
-    "rollup": "^0.59.4",
-    "rollup-plugin-babel": "^3.0.4",
+    "rollup": "^0.61.2",
+    "rollup-plugin-babel": "^3.0.5",
     "rollup-plugin-commonjs": "^9.1.3",
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-size-snapshot": "^0.4.0",
+    "rollup-plugin-size-snapshot": "^0.5.1",
     "rollup-plugin-strip": "^1.1.1",
-    "rollup-plugin-uglify": "^3.0.0",
-    "styled-components": "^3.3.2",
-    "stylelint": "9.2.1",
+    "rollup-plugin-uglify": "^4.0.0",
+    "styled-components": "^3.3.3",
+    "stylelint": "9.3.0",
     "stylelint-config-prettier": "^3.3.0",
     "stylelint-config-standard": "^18.2.0",
     "stylelint-config-styled-components": "^0.1.1",
